# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GaskeunKerja for Business \u2014 MVP",
    "version": "0.1.0"
  },
  "paths": {
...
    "/jobs": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Create Job",
        "operationId": "create_job_jobs_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "List Jobs",
        "operationId": "list_jobs_jobs_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobListItemOut"
                  },
                  "title": "Response List Jobs Jobs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/jobs/{job_id}/questions/generate": {
      "post": {
        "tags": [
          "interview_questions"
        ],
        "summary": "Generate Job Question",
        "description": "Stateless (2026-07-19 follow-up, user-requested): generates ONE question TEXT via Flash and\nreturns it WITHOUT writing anything to the database. AI-suggested drafts now live purely in\nfrontend state until HR clicks \"Setujui & Kunci Pertanyaan\" (PUT /questions then POST\n/questions/approve) \u2014 matching how manually-typed drafts (via \"+ Tambah Pertanyaan\") already\nonly ever get persisted at that point.\n\nGenerates exactly ONE question per call, not a batch \u2014 a single-prompt \"generate all N, stay\ndistinct\" attempt was tried first but the user found the real-world results still\ninconsistent/mixed/unrelated. The frontend now calls this endpoint once per requested question,\nIN ORDER, accumulating results into body.existing_questions on each subsequent call (so\ndistinctness is enforced by construction) \u2014 deliberately client-driven rather than looped\nserver-side so the UI can show live per-question progress. body.hint carries that specific\nslot's own current text (a real finding: discarding it and generating a generic JD-grounded\nquestion produced results with no connection to what HR had actually written for that slot).",
        "operationId": "generate_job_question_jobs__job_id__questions_generate_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateQuestionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratedQuestionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}/questions": {
      "get": {
        "tags": [
          "interview_questions"
        ],
        "summary": "List Job Questions",
        "operationId": "list_job_questions_jobs__job_id__questions_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuestionOut"
                  },
                  "title": "Response List Job Questions Jobs  Job Id  Questions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "interview_questions"
        ],
        "summary": "Update Job Questions",
        "description": "HR edits/adds/removes draft questions. Only draft-status questions can be edited.",
        "operationId": "update_job_questions_jobs__job_id__questions_put",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionsUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuestionOut"
                  },
                  "title": "Response Update Job Questions Jobs  Job Id  Questions Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}/questions/{question_id}/duration": {
      "patch": {
        "tags": [
          "interview_questions"
        ],
        "summary": "Update Question Duration",
        "description": "Per-question time limit (2026-07-19 follow-up, replacing the earlier job-wide-only design)\n\u2014 editable independent of draft/approved status, mirroring the job-level duration PATCH's\nimmediate-save pattern. Doesn't touch question_text/order_index/status, so it never needs a\nreopen even on an approved (locked) question.",
        "operationId": "update_question_duration_jobs__job_id__questions__question_id__duration_patch",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          },
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Question Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionDurationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}/questions/reopen": {
      "post": {
        "tags": [
          "interview_questions"
        ],
        "summary": "Reopen Job Questions",
        "description": "Flips already-approved questions back to draft so HR can add/edit before re-approving.",
        "operationId": "reopen_job_questions_jobs__job_id__questions_reopen_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuestionOut"
                  },
                  "title": "Response Reopen Job Questions Jobs  Job Id  Questions Reopen Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}/questions/approve": {
      "post": {
        "tags": [
          "interview_questions"
        ],
        "summary": "Approve Job Questions",
        "description": "Flips all draft questions to approved. Candidates only ever see approved questions.",
        "operationId": "approve_job_questions_jobs__job_id__questions_approve_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuestionOut"
                  },
                  "title": "Response Approve Job Questions Jobs  Job Id  Questions Approve Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GenerateQuestionRequest": {
        "properties": {
          "hint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hint"
          },
          "existing_questions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Existing Questions",
            "default": []
          }
        },
        "type": "object",
        "title": "GenerateQuestionRequest"
      },
      "GeneratedQuestionOut": {
        "properties": {
          "question": {
            "type": "string",
            "title": "Question"
          }
        },
        "type": "object",
        "required": [
          "question"
        ],
        "title": "GeneratedQuestionOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobCreateRequest": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "responsibilities": {
            "type": "string",
            "title": "Responsibilities"
          },
          "requirements": {
            "type": "string",
            "title": "Requirements"
          },
          "qualifications": {
            "type": "string",
            "title": "Qualifications"
          }
        },
        "type": "object",
        "required": [
          "title",
          "responsibilities",
          "requirements",
          "qualifications"
        ],
        "title": "JobCreateRequest"
      },
...
      "JobListItemOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "company_id": {
            "type": "integer",
            "title": "Company Id"
          },
          "company_name": {
            "type": "string",
            "title": "Company Name"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "responsibilities": {
            "type": "string",
            "title": "Responsibilities"
          },
          "requirements": {
            "type": "string",
            "title": "Requirements"
          },
          "qualifications": {
            "type": "string",
            "title": "Qualifications"
          },
          "required_education_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Required Education Level"
          },
          "required_major": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Required Major"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "candidate_count": {
            "type": "integer",
            "title": "Candidate Count"
          },
          "question_status": {
            "type": "string",
            "title": "Question Status"
          },
          "belum_diundang": {
            "type": "integer",
            "title": "Belum Diundang"
          },
          "menunggu_wawancara": {
            "type": "integer",
            "title": "Menunggu Wawancara"
          },
          "selesai_wawancara": {
            "type": "integer",
            "title": "Selesai Wawancara"
          },
          "diputuskan": {
            "type": "integer",
            "title": "Diputuskan"
          }
        },
        "type": "object",
        "required": [
          "id",
          "company_id",
          "company_name",
          "title",
          "responsibilities",
          "requirements",
          "qualifications",
          "required_education_level",
          "required_major",
          "status",
          "created_at",
          "candidate_count",
          "question_status",
          "belum_diundang",
          "menunggu_wawancara",
          "selesai_wawancara",
          "diputuskan"
        ],
        "title": "JobListItemOut"
      },
      "JobOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "company_id": {
            "type": "integer",
            "title": "Company Id"
          },
          "company_name": {
            "type": "string",
            "title": "Company Name"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "responsibilities": {
            "type": "string",
            "title": "Responsibilities"
          },
          "requirements": {
            "type": "string",
            "title": "Requirements"
          },
          "qualifications": {
            "type": "string",
            "title": "Qualifications"
          },
          "required_education_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Required Education Level"
          },
          "required_major": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Required Major"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "company_id",
          "company_name",
          "title",
          "responsibilities",
          "requirements",
          "qualifications",
          "required_education_level",
          "required_major",
          "status",
          "created_at"
        ],
        "title": "JobOut"
      },
...
      "QuestionDurationRequest": {
        "properties": {
          "duration_seconds": {
            "type": "integer",
            "title": "Duration Seconds"
          }
        },
        "type": "object",
        "required": [
          "duration_seconds"
        ],
        "title": "QuestionDurationRequest"
      },
      "QuestionOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "job_id": {
            "type": "integer",
            "title": "Job Id"
          },
          "question_text": {
            "type": "string",
            "title": "Question Text"
          },
          "order_index": {
            "type": "integer",
            "title": "Order Index"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "duration_seconds": {
            "type": "integer",
            "title": "Duration Seconds"
          }
        },
        "type": "object",
        "required": [
          "id",
          "job_id",
          "question_text",
          "order_index",
          "status",
          "duration_seconds"
        ],
        "title": "QuestionOut"
      },
      "QuestionUpdateItem": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "question_text": {
            "type": "string",
            "title": "Question Text"
          },
          "order_index": {
            "type": "integer",
            "title": "Order Index"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          }
        },
        "type": "object",
        "required": [
          "question_text",
          "order_index"
        ],
        "title": "QuestionUpdateItem"
      },
      "QuestionsUpdateRequest": {
        "properties": {
          "questions": {
            "items": {
              "$ref": "#/components/schemas/QuestionUpdateItem"
            },
            "type": "array",
            "title": "Questions"
          }
        },
        "type": "object",
        "required": [
          "questions"
        ],
        "title": "QuestionsUpdateRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}